+++ sql/vector--0.8.6--0.8.7.sql
@@ -243,6 +243,11 @@ COMMENT ON FUNCTION cosine_distance(sparsevec, sparsevec) IS 'cosine distance';
 
 COMMENT ON FUNCTION l1_distance(sparsevec, sparsevec) IS 'taxicab distance';
 
+CREATE FUNCTION vector_dims(sparsevec) RETURNS integer
+	AS 'MODULE_PATHNAME', 'sparsevec_vector_dims' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
+
+COMMENT ON FUNCTION vector_dims(sparsevec) IS 'number of dimensions';
+
 COMMENT ON FUNCTION l2_norm(sparsevec) IS 'Euclidean norm';
 
 COMMENT ON FUNCTION l2_normalize(sparsevec) IS 'normalize with Euclidean norm';